# Fix Unmarshal to used struct